image-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
rembg==2.0.50
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
